entities

+++ collective-blog/src/main/java/ru/barmatin/collectiveblog/repository/BlogUserRepository.java
@@ -3,7 +3,7 @@ package ru.barmatin.collectiveblog.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.barmatin.collectiveblog.domain.BlogUser;
 
-public interface BlogUserRepository extends JpaRepository<BlogUser, Integer> {
+public interface BlogUserRepository extends JpaRepository<BlogUser, Long> {
 
     BlogUser findByUsername(String username);
